170422-001

+++ src/main/java/com/banking/customer/service/app/service/imp/CustomerServiceImp.java
@@ -1,8 +1,10 @@
 package com.banking.customer.service.app.service.imp;
 
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.banking.customer.service.app.enity.DocData;
 import com.banking.customer.service.app.model.Customer;
 import com.banking.customer.service.app.repository.CustomerRepository;
 import com.banking.customer.service.app.service.CustomerService;
@@ -22,23 +24,73 @@ public class CustomerServiceImp implements CustomerService {
 	}
 
 	@Override
-	public Mono<Customer> findById(String id) {
-		return customerRepository.findById(id);
+	public Mono<Object> findById(String id) {
+		return customerRepository.findById(id)
+				.defaultIfEmpty(new Customer())
+				.flatMap(c -> {
+					if (c.getId() == null) {
+						return Mono.just(new DocData(id,"404","Customer Not Found"));
+					}
+					return Mono.just(c);
+				});
 	}
 
 	@Override
-	public Mono<Customer> findByPersonalIdentifier(int personalIdentifier) {
-		return null;
+	public Mono<Object> findByPersonalIdentifier(Integer personalIdentifier) {
+		
+		Mono<Object> dataObject = customerRepository.findByPersonalIdentifier(personalIdentifier)
+				.defaultIfEmpty(new Customer())
+				.flatMap(c -> {
+					if (c.getId() == null) {
+						return Mono.just(new DocData(null,"404","Customer Not Found",personalIdentifier.toString()));
+					}
+					return Mono.just(c);
+				});
+		
+		return dataObject;
 	}
 
 	@Override
-	public Mono<Customer> save(Customer customer) {
-		return customerRepository.save(customer);
+	public Mono<Object> save(Customer customer) {
+		
+		Mono<Object> customerData = customerRepository.findByPersonalIdentifier(customer.getPersonalIdentifier())
+				.defaultIfEmpty(new Customer())
+				.flatMap(c -> {
+					if (c.getId() != null) {
+						return Mono.just(new DocData(c.getId(),"302","Customer Already Exist",customer.getPersonalIdentifier().toString()));
+					}
+					return customerRepository.save(customer);
+				});
+	
+		return customerData;
 	}
 
 	@Override
-	public Mono<Void> delete(Customer customer) {
-		return customerRepository.delete(customer);
+	public Mono<Object> delete(Customer customer) {
+		
+		Mono<Object> customerData = customerRepository.findById(customer.getId())
+				.defaultIfEmpty(new Customer())
+				.flatMap(c -> {
+					if (c.getId() == null) {
+						return Mono.just(new DocData(c.getId(),"404","Inexisting Customer, can't be eliminated"));
+					}
+					return customerRepository.delete(customer)
+							.then(Mono.just(new DocData(c.getId(),"200","Correct Deletig",c.getPersonalIdentifier().toString())));
+				});
+		return customerData;
+	}
+
+	@Override
+	public Mono<Object> findByIdOrPersonalIdentifier(String id, Integer personalIdentifier) {
+		Mono<Object> customerData = customerRepository.findByIdOrPersonalIdentifier(id, personalIdentifier)
+				.defaultIfEmpty(new Customer())
+				.flatMap(c -> {
+					if (c.getId() != null) {
+						return Mono.just(new DocData(c.getId(),"404","Inexisting Customer, can't be eliminated",personalIdentifier.toString()));
+					}
+					return Mono.just(c);
+				});
+		return customerData;
 	}
 
 }
